requirements.txt
flask
qingstor-sdk
